smm/requirements.txt
Flask==2.1.0
vk-api==11.9.0
python-dotenv==0.19.1
gunicorn==20.1.0
python-telegram-bot==13.7
psycopg2-binary==2.9.1
